send post and navigat to chatroom changes done
